remove unnecessary PathProcessor interface
